pyproject.toml
[project]
name = "test-itk"
version = "0.0.1"
description = "Тестовое задание, реализовать API."
authors = [
    { name = "Никита Шидогубов", email = "anginike@ya.ru" },
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django-filter>=25.1",
    "djangorestframework>=3.16.1",
    "drf-spectacular>=0.28.0",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "faker>=37.8.0",
    "pre-commit>=4.3.0",
    "ruff>=0.13.1",
    "tqdm>=4.67.1",
]

[tool.ruff]
line-length = 119
exclude = [
    ".git",
    ".gitignore",
    "__pycache__",
    "__init__.py",
    "pyproject.toml",
    "tests.py",
]
target-version = "py313"
fix = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[tool.ruff.lint]
ignore = ["E501", "W293"]
select = ["E", "F", "I", "B", "C", "A", "N", "UP"]
mccabe.max-complexity = 12
